move write_posterior to base mcmc/multitemper classes
