+++ pycbc/inference/io/emcee_pt.py
@@ -21,7 +21,6 @@ from __future__ import absolute_import
 import h5py, numpy
 from .base_hdf import BaseInferenceFile
 from .base_multitemper import (MultiTemperedMetadataIO, MultiTemperedMCMCIO)
-from .posterior import PosteriorFile
 
 
 class EmceePTFile(MultiTemperedMCMCIO, MultiTemperedMetadataIO,
@@ -93,27 +92,3 @@ class EmceePTFile(MultiTemperedMCMCIO, MultiTemperedMetadataIO,
         except KeyError:
             # dataset doesn't exist yet, create it
             self[group] = acceptance_fraction
-
-    def write_posterior(self, filename, **kwargs):
-        """Write posterior only file
-
-        Parameters
-        ----------
-        filename : str
-            Name of output file to store posterior
-        """
-        f = h5py.File(filename, 'w')
-
-        # Preserve top-level metadata
-        for key in self.attrs:
-            f.attrs[key] = self.attrs[key]
-
-        f.attrs['filetype'] = PosteriorFile.name
-        s = f.create_group('samples')
-        fields = self[self.samples_group].keys()
-
-        # Copy and squash fields into one dimensional arrays
-        for field_name in fields:
-            fvalue = self[self.samples_group][field_name][:]
-            thin = fvalue[0,:,self.thin_start:self.thin_end:self.thin_interval]
-            s[field_name] = thin.flatten()
