+++ pycbc/inference/io/ptemcee.py
@@ -20,7 +20,6 @@
 from __future__ import absolute_import
 from .base_hdf import BaseInferenceFile
 from .base_multitemper import (MultiTemperedMetadataIO, MultiTemperedMCMCIO)
-from .posterior import PosteriorFile
 
 
 class PTEmceeFile(MultiTemperedMCMCIO, MultiTemperedMetadataIO,
@@ -132,27 +131,3 @@ class PTEmceeFile(MultiTemperedMCMCIO, MultiTemperedMetadataIO,
         """
         group = self[self.sampler_group]
         return {attr: group[attr][:] for attr in self._ensemble_attrs}
-
-    def write_posterior(self, filename, **kwargs):
-        """Write posterior only file
-
-        Parameters
-        ----------
-        filename : str
-            Name of output file to store posterior
-        """
-        f = h5py.File(filename, 'w')
-
-        # Preserve top-level metadata
-        for key in self.attrs:
-            f.attrs[key] = self.attrs[key]
-
-        f.attrs['filetype'] = PosteriorFile.name
-        s = f.create_group('samples')
-        fields = self[self.samples_group].keys()
-
-        # Copy and squash fields into one dimensional arrays
-        for field_name in fields:
-            fvalue = self[self.samples_group][field_name][:]
-            thin = fvalue[0,:,self.thin_start:self.thin_end:self.thin_interval]
-            s[field_name] = thin.flatten()
